+++ src/main/java/vn/kms/ngaythobet/domain/tournament/RoundService.java
@@ -53,9 +53,12 @@ public class RoundService {
     }
 
     public void createRound(CreateRoundInfo createRoundInfo) {
+        if (createRoundInfo.getName().trim().length() < 6) {
+            throw new DataInvalidException("validation.name.notEmpty.size.blankspace");
+        }
         if (!roundNameIsExistedInTournament(createRoundInfo.getName(), createRoundInfo.getTournamentId())) {
             Round round = new Round();
-            round.setName(createRoundInfo.getName());
+            round.setName(createRoundInfo.getName().trim());
             Tournament tournament = tournamentRepo.getOne(createRoundInfo.getTournamentId());
             round.setTournament(tournament);
             if (createRoundInfo.getCompetitorIds() != null) {
@@ -97,6 +100,11 @@ public class RoundService {
     @Transactional(readOnly = true)
     public List<Round> getRoundByTournamentId(Long tournamentId) {
         return roundRepo.findByTournamentId(tournamentId);
+    }
+
+    @Transactional(readOnly = true)
+    public Round findByBettingMatchId(Long bettingMatchId) {
+        return roundRepo.findByBettingMatchId(bettingMatchId);
     }
 
 }