+++ src/main/java/vn/kms/ngaythobet/domain/tournament/Tournament.java
@@ -120,6 +120,9 @@ public class Tournament extends AuditableEntity {
 
     @Override
     public int hashCode() {
+        if (this.getId() == null) {
+            return ObjectUtils.hashCode((long) 0);
+        }
         return ObjectUtils.hashCode(this.getId());
     }