+++ src/main/java/vn/kms/ngaythobet/domain/tournament/TournamentService.java
@@ -1,6 +1,7 @@
 // Copyright (c) 2015 KMS Technology, Inc.
 package vn.kms.ngaythobet.domain.tournament;
 
+import java.util.ArrayList;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -16,6 +17,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import vn.kms.ngaythobet.domain.core.User;
 import vn.kms.ngaythobet.domain.core.UserRepository;
+import vn.kms.ngaythobet.domain.util.DataInvalidException;
 import vn.kms.ngaythobet.domain.util.Constants;
 import vn.kms.ngaythobet.domain.util.DataInvalidException;
 import vn.kms.ngaythobet.domain.util.SecurityUtil;
@@ -31,6 +33,12 @@ public class TournamentService {
     private String UPLOAD_FILE_LOCATION;
 
     @Autowired
+    private RoundRepository roundRepo;
+
+    @Autowired
+    private MatchRepository matchRepo;
+
+    @Autowired
     private UserRepository userRepo;
 
     @Autowired
@@ -40,8 +48,11 @@ public class TournamentService {
     }
 
     public void createTournament(CreateTournamentInfo tournamentInfo) {
+        if (tournamentInfo.getName().trim().length() < 6) {
+            throw new DataInvalidException("validation.name.notEmpty.size.blankspace");
+        }
         Tournament tournament = new Tournament();
-        tournament.setName(tournamentInfo.getName());
+        tournament.setName(tournamentInfo.getName().trim());
         tournament.setNumOfCompetitor((long) tournamentInfo.getCompetitors().size());
         tournament.setActivated(tournamentInfo.isActive());
         tournamentRepo.save(tournament);
@@ -81,6 +92,46 @@ public class TournamentService {
             return tournaments;
         }
         return tournamentRepo.findAll();
+    }
+
+    public void saveTournament(int index) {
+        List<Tournament> tournaments = ParseData.parseTournamentFromLiveScore();
+        Tournament tournamentTest = tournaments.get(index);
+        tournamentRepo.save(tournamentTest);
+        Tournament tournament = tournamentRepo.findByName(tournamentTest.getName());
+        List<Competitor> competitors = tournamentTest.getCompetitors();
+        for (Competitor competitor : competitors) {
+            competitor.setTournament(tournament);
+            competitorRepo.save(competitor);
+        }
+        for (Round round : tournamentTest.getRounds()) {
+            List<Competitor> rawCompetitors = round.getCompetitors();
+            List<Competitor> roundCompetitors = new ArrayList<>();
+
+            for (Competitor competitor : rawCompetitors) {
+                Competitor roundCompetior = competitorRepo.findByTournamentAndName(tournament, competitor.getName());
+                List<Round> rounds = roundCompetior.getRounds();
+                if (rounds == null) {
+                    rounds = new ArrayList<>();
+                }
+                rounds.add(round);
+                roundCompetior.setRounds(rounds);
+                roundCompetitors.add(roundCompetior);
+            }
+            round.setCompetitors(roundCompetitors);
+            round.setTournament(tournamentTest);
+            roundRepo.save(round);
+            for (Match match : round.getMatches()) {
+                Competitor competitor1 = competitorRepo.findByTournamentAndName(tournament,
+                        match.getCompetitor1().getName());
+                Competitor competitor2 = competitorRepo.findByTournamentAndName(tournament,
+                        match.getCompetitor2().getName());
+                match.setCompetitor1(competitor1);
+                match.setCompetitor2(competitor2);
+                match.setRound(roundRepo.findByNameAndTournament(round.getName(), tournament));
+                matchRepo.save(match);
+            }
+        }
     }
 
     public void uploadTournmentImage(MultipartFile file, Long tournamentId) {