+++ src/test/java/vn/kms/ngaythobet/domain/betting/BettingMatchServiceTest.java
@@ -151,7 +151,7 @@ public class BettingMatchServiceTest extends BaseTest {
         createBettingMatchInfo.setBalance2(new BigDecimal("0.5"));
         createBettingMatchInfo.setBetAmount(new BigDecimal("1000"));
         createBettingMatchInfo.setDecription("test");
-        createBettingMatchInfo.setExpiredTime(LocalDateTime.now());
+        createBettingMatchInfo.setExpiredTime(LocalDateTime.now().plusDays(1));
         createBettingMatchInfo.setGroupId(groupTemp1.getId());
         createBettingMatchInfo.setMatchId(matchTemp.getId());
         bettingMatchService.createBettingMatch(createBettingMatchInfo);
@@ -167,12 +167,27 @@ public class BettingMatchServiceTest extends BaseTest {
         createBettingMatchInfo.setBalance2(new BigDecimal("0.5"));
         createBettingMatchInfo.setBetAmount(new BigDecimal("1000"));
         createBettingMatchInfo.setDecription("test");
-        createBettingMatchInfo.setExpiredTime(LocalDateTime.now());
+        createBettingMatchInfo.setExpiredTime(LocalDateTime.now().plusDays(1));
         createBettingMatchInfo.setGroupId(groupTemp2.getId());
         createBettingMatchInfo.setMatchId(matchTemp.getId());
         exception.expectMessage("{exception.match.group.invalid}");
         bettingMatchService.createBettingMatch(createBettingMatchInfo);
-        ;
+        assertThat(bettingMatchRepo.findAll().size(), equalTo(1));
+    }
+    
+    @Test
+    public void testCreateBettingMatchWithInvalidExpiredTime() {
+        CreateBettingMatchInfo createBettingMatchInfo = new CreateBettingMatchInfo();
+        createBettingMatchInfo.setActivated(true);
+        createBettingMatchInfo.setBalance1(new BigDecimal("0"));
+        createBettingMatchInfo.setBalance2(new BigDecimal("0.5"));
+        createBettingMatchInfo.setBetAmount(new BigDecimal("1000"));
+        createBettingMatchInfo.setDecription("test");
+        createBettingMatchInfo.setExpiredTime(LocalDateTime.now());
+        createBettingMatchInfo.setGroupId(groupTemp1.getId());
+        createBettingMatchInfo.setMatchId(matchTemp.getId());
+        exception.expectMessage("{exception.expiredTime.invalid}");
+        bettingMatchService.createBettingMatch(createBettingMatchInfo);
         assertThat(bettingMatchRepo.findAll().size(), equalTo(1));
     }
 
@@ -184,7 +199,7 @@ public class BettingMatchServiceTest extends BaseTest {
         createBettingMatchInfo.setBalance2(new BigDecimal("0.1111"));
         createBettingMatchInfo.setBetAmount(new BigDecimal("1000"));
         createBettingMatchInfo.setDecription("test");
-        createBettingMatchInfo.setExpiredTime(LocalDateTime.now());
+        createBettingMatchInfo.setExpiredTime(LocalDateTime.now().plusDays(1l));
         createBettingMatchInfo.setGroupId(groupTemp2.getId());
         createBettingMatchInfo.setMatchId(matchTemp.getId());
         exception.expectMessage("{exception.balance.in.valid}");
@@ -207,7 +222,6 @@ public class BettingMatchServiceTest extends BaseTest {
         createBettingMatchInfo.setMatchId(bettingMatch.getMatch().getId());
         exception.expectMessage("{exception.bettingMatch.is.existed}");
         bettingMatchService.createBettingMatch(createBettingMatchInfo);
-        ;
         assertThat(bettingMatchRepo.findAll().size(), equalTo(1));
     }
 
@@ -242,7 +256,7 @@ public class BettingMatchServiceTest extends BaseTest {
         updateBettingMatchInfo.setBetAmount(new BigDecimal("999999"));
         updateBettingMatchInfo.setBettingMatchId(bettingMatch.getId());
         updateBettingMatchInfo.setDecription("test again");
-        updateBettingMatchInfo.setExpiredTime(LocalDateTime.now().plusDays(1));
+        updateBettingMatchInfo.setExpiredTime(LocalDateTime.now().plusDays(2));
         updateBettingMatchInfo.setGroupId(groupTemp2.getId());
         updateBettingMatchInfo.setMatchId(matchTemp.getId());
         exception.expectMessage("{exception.match.group.invalid}");
@@ -253,7 +267,6 @@ public class BettingMatchServiceTest extends BaseTest {
     @Test
     public void testUpdateBettingMatchWithInvalidBalance() {
         BettingMatch bettingMatch = createBettingMatch();
-
         UpdateBettingMatchInfo updateBettingMatchInfo = new UpdateBettingMatchInfo();
         updateBettingMatchInfo.setBalance1(new BigDecimal(1.111111));
         updateBettingMatchInfo.setBalance2(new BigDecimal(0));
@@ -265,6 +278,24 @@ public class BettingMatchServiceTest extends BaseTest {
         updateBettingMatchInfo.setGroupId(groupTemp2.getId());
         updateBettingMatchInfo.setMatchId(matchTemp.getId());
         exception.expectMessage("{exception.balance.in.valid}");
+        bettingMatchService.updateBettingMatch(updateBettingMatchInfo);
+        assertThat(bettingMatchRepo.findAll().size(), equalTo(1));
+    }
+    
+    @Test
+    public void testUpdateBettingMatchWithInvalidExpiredTime() {
+        BettingMatch bettingMatch = createBettingMatch();
+        UpdateBettingMatchInfo updateBettingMatchInfo = new UpdateBettingMatchInfo();
+        updateBettingMatchInfo.setBalance1(new BigDecimal(1.111111));
+        updateBettingMatchInfo.setBalance2(new BigDecimal(0));
+        updateBettingMatchInfo.setActivated(false);
+        updateBettingMatchInfo.setBetAmount(new BigDecimal("999999"));
+        updateBettingMatchInfo.setBettingMatchId(bettingMatch.getId());
+        updateBettingMatchInfo.setDecription("test again");
+        updateBettingMatchInfo.setExpiredTime(LocalDateTime.now());
+        updateBettingMatchInfo.setGroupId(groupTemp2.getId());
+        updateBettingMatchInfo.setMatchId(matchTemp.getId());
+        exception.expectMessage("{exception.expiredTime.invalid}");
         bettingMatchService.updateBettingMatch(updateBettingMatchInfo);
         assertThat(bettingMatchRepo.findAll().size(), equalTo(1));
     }