+++ CHANGELOG.md
@@ -1,13 +1,29 @@
+ngaythobet - v1.0.3 / 2016-01-13
+================================
+**Improvement**
+* [UI] Click on link breadscrumb
+* [UI] Bookmark URL
+* [API] Create auto tournament from livescore.com
+* [API] Check round when create match
+* [API] Add image into tournament
+ 
+
+**Bug Fixed** 
+* [#141] System allows to create round with Name = 'null'
+* [#143] System doesn't display competitor for default round (round 'Circle' ) in create new match if this tournament has another round.
+* [#144] System can't show error msg when creating new match with all field is empty.
+
+
 ngaythobet - v1.0.2 / 2016-01-12
 ================================
 **Improvement**
 * [UI] Update tournament UI
 * [UI] Completed comment, translate language
 * [API] Updated role ADMIN can view any group statistic
+* [API] Group statistic - udpate getComment 
 
 **Bug Fixed** 
 * [#122] Player can view ALL group in Tournament when they just join only group in this tournament
-* [#136[ System should not displayed 'add competitor' field in update round when all competitor have been selected in round
 
 
 ngaythobet - v1.0.1 / 2016-01-11
@@ -25,6 +41,7 @@ ngaythobet - v1.0.1 / 2016-01-11
 * [#131] Team's score displays null value when user selects that match to update score
 * [#132] Total loss amount is incremental when user clicks on betting group that user viewing
 * [#133] System translate from EN to VI language is not correct
+* [#136[ System should not displayed 'add competitor' field in update round when all competitor have been selected in round
 
 
 ngaythobet - v1.0.0 / 2016-01-08